Update image-dialog.component HTML and SCSS
Updated the HTML template of the `image-dialog.component` to include a new `div` with the class `selectedImage`, containing an `img` element with dynamic `src` and alternate text. Added a `button-container` `div` with delete and download buttons, conditional classes, and a date display. Included a GIF image and a close button that triggers the `onClose()` method. Corrected the import path in the SCSS file for `Variables.scss`.