Add logic to set DateTaken for AddedImages in MolenService
Added a new block of logic in MolenService.cs to check if the
AddedImages collection of each MolenData item contains any images.
If images are present, it iterates through each AddedImage and
sets the DateTaken property using the GetDateTaken method if
DateTaken is null. The AddedImage is then updated in the database
asynchronously using _db.UpdateAsync(addedImg).