Refactor data models to use "Old" versions
Updated project structure to include new folders for migrations and models. Modified `DbConnection.cs` to create new tables with the suffix "Old". Refactored methods in `MolenService.cs` and `PlacesService.cs` to work with the new "Old" data models, including changes to data retrieval and handling of images. This transition indicates a versioning strategy in the database schema.